requirements.txt
datasets==2.19.1
openai==1.30.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.40.2
vllm==0.4.2
anthropic==0.34.2
google.generativeai==0.8.2
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=1.5.0
seaborn>=0.11.0
scipy>=1.9.0
PyYAML>=6.0
